scanner-proxy/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pillow==11.3.0
